requirements.txt
crochet==1.12.0
Django==3.1.7
django-mysql==3.11.1
djangorestframework==3.12.4
google==3.0.0
python-dotenv==0.16.0
Scrapy
